requirements.txt
amqp==2.5.2
appnope==0.1.0
asgiref==3.2.3
backcall==0.1.0
beautifulsoup4==4.8.2
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
chardet==3.0.4
cuid==0.3
decorator==4.4.2
Django==3.0.3
django-celery-beat==2.0.0
django-cors-headers==3.2.1
django-timezone-field==4.0
djangorestframework==3.11.0
feedparser==5.2.1
idna==2.9
ipdb==0.13.2
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
joblib==0.14.1
kombu==4.6.8
numpy==1.18.1
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.4
ptyprocess==0.6.0
Pygments==2.6.1
python-crontab==2.4.2
python-dateutil==2.8.1
python-opengraph==0.0.2
pytz==2019.3
redis==3.4.1
requests==2.22.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
traitlets==4.3.3
urlcanon==0.3.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
